[refactor] : 접근제한자 수정 및 폴더 추가
